requirements.txt
gymnasium==0.29.1
numpy==1.24.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
scikit-learn==1.3.2
